11091: Auto-correct all existing cops (safe)

+++ db/migrate/20180103201901_remove_uuid_cols.rb
@@ -1,9 +1,9 @@
+# frozen_string_literal: true
+
 class RemoveUuidCols < ActiveRecord::Migration[4.2]
   def up
     ActiveRecord::Base.connection.tables.each do |table|
-      if column_exists?(table, "uuid")
-        remove_column table, "uuid"
-      end
+      remove_column table, "uuid" if column_exists?(table, "uuid")
     end
   end
